Move bflsc defines to a header file.
